[gym/common] Add 'compute_tilt_from_quat', 'swing_from_vector', and 'remove_twist_from_quat' utilities.
