+++ python/gym_jiminy/common/gym_jiminy/common/blocks/mahony_filter.py
@@ -2,7 +2,7 @@
 reinforcement learning pipeline environment design.
 """
 import logging
-from typing import List, Union, Optional, Tuple, no_type_check
+from typing import List, Union, Optional
 
 import numpy as np
 import numba as nb
@@ -12,30 +12,16 @@ from jiminy_py.core import (  # pylint: disable=no-name-in-module
     ImuSensor as imu)
 
 from ..bases import BaseObsT, BaseActT, BaseObserverBlock, InterfaceJiminyEnv
-from ..utils import ArrayOrScalar, fill, matrices_to_quat
+from ..utils import (fill,
+                     matrices_to_quat,
+                     swing_from_vector,
+                     compute_tilt_from_quat,
+                     remove_twist_from_quat)
 
 
 EARTH_SURFACE_GRAVITY = 9.81
-TWIST_SWING_SINGULARITY_THR = 1e-6
 
 LOGGER = logging.getLogger(__name__)
-
-
-@nb.jit(nopython=True, cache=True)
-def compute_tilt(q: np.ndarray) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
-    """Compute e_z in R(q) frame (Euler-Rodrigues Formula): R(q).T @ e_z.
-
-    :param q: Array whose rows are the 4 components of quaternions (x, y, z, w)
-              and columns are N independent orientations.
-
-    :returns: Tuple of arrays corresponding to the 3 individual components
-              (a_x, a_y, a_z) of N independent tilt axes.
-    """
-    q_x, q_y, q_z, q_w = q
-    v_x = 2 * (q_x * q_z - q_y * q_w)
-    v_y = 2 * (q_y * q_z + q_w * q_x)
-    v_z = 1 - 2 * (q_x * q_x + q_y * q_y)
-    return (v_x, v_y, v_z)
 
 
 @nb.jit(nopython=True, nogil=True, cache=True)
@@ -71,7 +57,7 @@ def mahony_filter(q: np.ndarray,
 
     """
     # Compute expected Earth's gravity (Euler-Rodrigues Formula): R(q).T @ e_z
-    v_x, v_y, v_z = compute_tilt(q)
+    v_x, v_y, v_z = compute_tilt_from_quat(q)
 
     # Compute the angular velocity using Explicit Complementary Filter:
     # omega_mes = v_a_hat x v_a, where x is the cross product.
@@ -106,83 +92,6 @@ def mahony_filter(q: np.ndarray,
 
     # Update Gyro bias
     bias_hat -= dt * ki * omega_mes
-
-
-# FIXME: Enabling cache causes segfault on Apple Silicon
-@no_type_check
-@nb.jit(nopython=True, cache=False)
-def swing_from_vector(
-        v_a: Tuple[ArrayOrScalar, ArrayOrScalar, ArrayOrScalar],
-        q: np.ndarray) -> None:
-    """Compute the "smallest" rotation transforming vector 'v_a' in 'e_z'.
-
-    :param v_a: Tuple of arrays corresponding to the 3 individual components
-                (a_x, a_y, a_z) of N independent tilt axes.
-    :param q: Array where the result will be stored. The rows are the 4
-              components of quaternions (x, y, z, w) and columns are the N
-              independent orientations.
-    """
-    # Extract individual tilt components
-    v_x, v_y, v_z = v_a
-
-    # There is a singularity when the rotation axis of orientation estimate
-    # and z-axis are nearly opposites, i.e. v_z ~= -1. One solution that
-    # ensure continuity of q_w is picked arbitrarily using SVD decomposition.
-    # See `Eigen::Quaternion::FromTwoVectors` implementation for details.
-    if q.ndim > 1:
-        is_singular = np.any(v_z < -1.0 + TWIST_SWING_SINGULARITY_THR)
-    else:
-        is_singular = v_z < -1.0 + TWIST_SWING_SINGULARITY_THR
-    if is_singular:
-        if q.ndim > 1:
-            for i, q_i in enumerate(q.T):
-                swing_from_vector((v_x[i], v_y[i], v_z[i]), q_i)
-        else:
-            _, _, v_h = np.linalg.svd(np.array((
-                (v_x, v_y, v_z),
-                (0.0, 0.0, 1.0))
-            ), full_matrices=True)
-            w_2 = (1 + max(v_z, -1)) / 2
-            q[:3], q[3] = v_h[-1] * np.sqrt(1 - w_2), np.sqrt(w_2)
-    else:
-        s = np.sqrt(2 * (1 + v_z))
-        q[0], q[1], q[2], q[3] = v_y / s, - v_x / s, 0.0, s / 2
-
-    # First order quaternion normalization to prevent compounding of errors.
-    # If not done, shit may happen with removing twist again and again on the
-    # same quaternion, which is typically the case when the IMU is steady, so
-    # that the mahony filter update is actually skipped internally.
-    q *= (3.0 - np.sum(np.square(q), 0)) / 2
-
-
-# FIXME: Enabling cache causes segfault on Apple Silicon
-@nb.jit(nopython=True, cache=False)
-def remove_twist(q: np.ndarray) -> None:
-    """Remove the twist part of the Twist-after-Swing decomposition of given
-    orientations in quaternion representation.
-
-    Any rotation R can be decomposed as:
-
-        R = R_z * R_s
-
-    where R_z (the twist) is a rotation around e_z and R_s (the swing) is
-    the "smallest" rotation matrix such that t(R_s) = t(R).
-
-    .. seealso::
-        * See "Estimation and control of the deformations of an exoskeleton
-          using inertial sensors", PhD Thesis, M. Vigne, 2021, p. 130.
-        * See "Swing-twist decomposition in Clifford algebra", P. Dobrowolski,
-          2015 (https://arxiv.org/abs/1506.05481)
-
-    :param q: Array whose rows are the 4 components of quaternions (x, y, z, w)
-              and columns are N independent orientations from which to remove
-              the swing part. It will be updated in-place.
-    """
-    # Compute e_z in R(q) frame (Euler-Rodrigues Formula): R(q).T @ e_z
-    v_a = compute_tilt(q)
-
-    # Compute the "smallest" rotation transforming vector 'v_a' in 'e_z'
-    swing_from_vector(v_a, q)
 
 
 @nb.jit(nopython=True, nogil=True, cache=True)
@@ -269,8 +178,8 @@ class MahonyFilter(
             integrator used to estimate the twist part of twist-after-swing
             decomposition of the estimated orientation in place of the Mahony
             Filter. If `0.0`, then its is kept constant equal to zero. `None`
-            to kept the original estimate provided by Mahony Filter.
-            See `remove_twist` and `update_twist` documentation for details.
+            to kept the original estimate provided by Mahony Filter. See
+            `remove_twist_from_quat` and `update_twist` doc for details.
             Optional: `0.0` by default.
         :param exact_init: Whether to initialize orientation estimate using
                            accelerometer measurements or ground truth. `False`
@@ -470,7 +379,7 @@ class MahonyFilter(
 
         # Remove twist if requested
         if self._remove_twist:
-            remove_twist(self.observation)
+            remove_twist_from_quat(self.observation)
 
         if self._update_twist:
             update_twist(self.observation,
