+++ python/gym_jiminy/common/gym_jiminy/common/utils/__init__.py
@@ -12,7 +12,10 @@ from .math import (squared_norm_2,
                    quat_multiply,
                    quat_average,
                    rpy_to_matrix,
-                   rpy_to_quat)
+                   rpy_to_quat,
+                   compute_tilt_from_quat,
+                   swing_from_vector,
+                   remove_twist_from_quat)
 from .spaces import (DataNested,
                      FieldNested,
                      ArrayOrScalar,
@@ -39,19 +42,22 @@ from .pipeline import build_pipeline, load_pipeline
 
 
 __all__ = [
-    "squared_norm_2",
-    "matrix_to_quat",
-    "matrices_to_quat",
-    "matrix_to_rpy",
-    "matrix_to_yaw",
-    "quat_to_matrix",
-    "quat_to_rpy",
-    "quat_to_yaw",
-    "quat_to_yaw_cos_sin",
-    "quat_multiply",
-    "quat_average",
-    "rpy_to_matrix",
-    "rpy_to_quat",
+    'squared_norm_2',
+    'matrix_to_quat',
+    'matrices_to_quat',
+    'matrix_to_rpy',
+    'matrix_to_yaw',
+    'quat_to_matrix',
+    'quat_to_rpy',
+    'quat_to_yaw',
+    'quat_to_yaw_cos_sin',
+    'quat_multiply',
+    'quat_average',
+    'rpy_to_matrix',
+    'rpy_to_quat',
+    'compute_tilt_from_quat',
+    'swing_from_vector',
+    'remove_twist_from_quat',
     'DataNested',
     'FieldNested',
     'ArrayOrScalar',
