+++ python/gym_jiminy/toolbox/gym_jiminy/toolbox/math/__init__.py
@@ -1,12 +1,10 @@
 # pylint: disable=missing-module-docstring
 
 from .qhull import ConvexHull, compute_distance_convex_to_point
-from .generic import quat_average
 
 __all__ = [
     "ConvexHull",
-    "compute_distance_convex_to_point",
-    "quat_average"
+    "compute_distance_convex_to_point"
 ]
 
 try:
